feat: expose autofirma configuration to HM module
These changes implement a new `programs.autofirma.config` submodule to configure AutoFirma via Home Manager.

To achieve this in an automated way, several new components are in place:
- **XMLDoclet**: A Java plugin included in `pom.xml` to export Javadoc in XML format.
- **properties-to-json**: A custom Python script that combines the information from the aforementioned Javadoc and a properties file present in the source code. This file contains the parameters that can be set in the AutoFirma configuration.

During the compilation process, the Javadoc is generated and then transformed into a JSON file containing a list of configuration settings with their default values and descriptions.

The Home Manager module reads this JSON file and builds the `config` submodule.

If the user modifies any of the values under `config`, a configuration file is generated and linked to the path that AutoFirma expects. To prevent AutoFirma from overwriting the configuration (which happens every time the configuration dialog is opened), the directory containing the link has its permissions modified to forbid the user from writing.