feat: externalize dependency hashes to a JSON file
This commit introduces a new tool `update-dependency-hashes` that
is accessible via `nix develop`.

The goal of this new tool is to recalculate the fixed-output derivation
hashes of the maven-dependencies (m2 repositories containing the jar
files needed to build AutoFirma.)

When this tool runs, it reads the (new) file `dependency-hashes.json` and
infer the order of execution, performing a topological sort of the
declared dependencies. Following that order, it goes through the
derivations one by one, clearing the current hash and rebuilding the
derivation. If a new hash is found, it updates it; otherwise, the hash
remains unchanged.